+++ OUDS/Core/Components/Sources/Internal/ControlItem/ControlItemBackgroudModifier.swift
@@ -1,0 +1,38 @@
+//
+// Software Name: OUDS iOS
+// SPDX-FileCopyrightText: Copyright (c) Orange SA
+// SPDX-License-Identifier: MIT
+//
+// This software is distributed under the MIT license,
+// the text of which is available at https://opensource.org/license/MIT/
+// or see the "LICENSE" file for more details.
+//
+// Authors: See CONTRIBUTORS.txt
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System
+//
+
+import OUDS
+import OUDSFoundations
+import OUDSTokensSemantic
+import SwiftUI
+
+/// A `ViewModifier` to apply to `ControlItem` views so as to define the background
+struct ControlItemBackgroundModifier: ViewModifier {
+
+    // MARK: Stored properties
+
+    let interactionState: InteractionState
+    @Environment(\.theme) private var theme
+
+    // MARK: Body
+
+    func body(content: Content) -> some View {
+        if interactionState == .hover {
+            content.oudsBackground(theme.controlItem.controlItemColorBgHover)
+        } else if interactionState == .pressed {
+            content.oudsBackground(theme.controlItem.controlItemColorBgPressed)
+        } else {
+            content
+        }
+    }
+}